requirements.txt
pymongo
python-dotenv
pandas
uuid
